// src/lib/types/api.ts
export interface Researcher {
  id: number
  name: string
  email: string
  grade: string
  teamId?: number
  role: "admin" | "researcher"
}
// ...
export interface StatsByYearItem {
  year: number
  count: number
}

export interface StatsByGroupItem {
  id: number
  name: string
  count: number
}